Income and Expense models for more advanced features (#23)

+++ db/migrate/20230810192224_create_accounts.rb
@@ -3,7 +3,7 @@ class CreateAccounts < ActiveRecord::Migration[7.0]
     create_table :accounts do |t|
       t.string :name, null: false
       t.string :account_type, null: false
-      t.integer :balance, null: false
+      t.decimal :balance, null: false
       t.references :user, null: false
 
       t.timestamps
